# Conflicts:
#	jplag/src/main/java/de/jplag/JPlagComparison.java
#	jplag/src/main/java/de/jplag/clustering/ClusteringAdapter.java
#	jplag/src/test/java/de/jplag/NormalComparisonTest.java
#	jplag/src/test/java/de/jplag/ParallelComparisonTest.java

+++ jplag/src/main/java/de/jplag/JPlagComparison.java
@@ -32,21 +32,21 @@ public record JPlagComparison(Submission firstSubmission, Submission secondSubmi
     /**
      * @return Maximum similarity in percent of both submissions.
      */
-    public final float maximalSimilarity() {
+    public final double maximalSimilarity() {
         return Math.max(similarityOfFirst(), similarityOfSecond());
     }
 
     /**
      * @return Minimum similarity in percent of both submissions.
      */
-    public final float minimalSimilarity() {
+    public final double minimalSimilarity() {
         return Math.min(similarityOfFirst(), similarityOfSecond());
     }
 
     /**
      * @return Similarity in percent (what percentage of tokens across both submissions are matched).
      */
-    public final float similarity() {
+    public final double similarity() {
         boolean subtractBaseCode = firstSubmission.hasBaseCodeMatches() && secondSubmission.hasBaseCodeMatches();
         int divisorA = firstSubmission.getSimilarityDivisor(subtractBaseCode);
         int divisorB = secondSubmission.getSimilarityDivisor(subtractBaseCode);
@@ -57,7 +57,7 @@ public record JPlagComparison(Submission firstSubmission, Submission secondSubmi
      * @return Similarity in percent for the first submission (what percent of the first submission is similar to the
      * second).
      */
-    public final float similarityOfFirst() {
+    public final double similarityOfFirst() {
         int divisor = firstSubmission.getSimilarityDivisor(true);
         return similarity(divisor);
     }
@@ -66,7 +66,7 @@ public record JPlagComparison(Submission firstSubmission, Submission secondSubmi
      * @return Similarity in percent for the second submission (what percent of the second submission is similar to the
      * first).
      */
-    public final float similarityOfSecond() {
+    public final double similarityOfSecond() {
         int divisor = secondSubmission.getSimilarityDivisor(true);
         return similarity(divisor);
     }
@@ -76,7 +76,7 @@ public record JPlagComparison(Submission firstSubmission, Submission secondSubmi
         return firstSubmission.getName() + " <-> " + secondSubmission.getName();
     }
 
-    private float similarity(int divisor) {
-        return (divisor == 0 ? 0f : (getNumberOfMatchedTokens() * 100 / (float) divisor));
+    private double similarity(int divisor) {
+        return (divisor == 0 ? 0.0 : (getNumberOfMatchedTokens() * 100 / (double) divisor));
     }
 }
